wheeler: Rename and rework --entrypoints-dir arg to --entrypoints-path
Instead of having --entrypoints-dir as a separate function, always
generate entrypoints (if any) into the target dir, by default into
subdir <prefix>/bin/, which can be modified with --entrypoints-path,
i.e. like --modules-path works. This has been clarified in the help
text.

--expect-entrypoints can still be used to ensure that certain
entrypoints are defined in the wheel, and to only generate these.

Signed-off-by: Philip Eklöf <phi@pxy.se>